This change set improves performance of network transfers over the virtio-net network interface. As an example, the Warp (https://github.com/minio/warp) mixed benchamrk (run with `./warp mixed --autoterm --objects 40 --duration 1m`) against a minIO server showed a change in results from `Cluster Total: 59.45 MiB/s, 9.88 obj/s over 53s` to `Cluster Total: 83.11 MiB/s, 13.88 obj/s over 50s`.
The first commit increases the TCP send buffer size and TCP window size (since lwIP implements TCP window scaling, the window size is not limited to a maximum of 65535).
The second commit adds support for the VIRTIO_F_RING_EVENT_IDX feature, which allows a virtIO driver to indicate to the device when it wants to be notified of an update to used buffers, and similarly allows a virtIO device to indicate to the driver when it wants to be notified of an update to available buffers. By avoiding unneeded device interrupts and driver notifications, the number of VM exits and hypervisor syscalls (e.g. to execute KVM_SIGNAL_MSI ioctls) is minimized, which decreases CPU load on the host.
The third commit implements polling mode in the virtIO code and enables it in the TX queue of the virtio_net driver. When a virtIO queue is in polling mode, the driver does not want to be notified (via an interrupt) when a virtIO buffer in the queue has been used by the device. This has been shown to improve the speed of TCP transfers: for example, the average performance of a Go web server as measured by the Jenkins CI job increases by ~12%.